Move RouteComponent to wigdets folder.
