+++ src/app/widgets/index.ts
@@ -4,6 +4,7 @@ export * from './alert/alert.component';
 export * from './form';
 export * from './spinner';
 export * from './utils';
+export * from './link';
 export { DraalDataTableModule } from './datatables/datatables.module';
 export { DraalFormsModule } from './form/form.module';
 export { DraalAlertModule } from './alert/alert.module';
