+++ src/app/widgets/link/route.component.ts
@@ -1,0 +1,21 @@
+import { Component } from '@angular/core';
+
+
+export interface RouteComponentData {
+    link: Array<any>;
+    text: string;
+}
+
+@Component({
+    selector: 'dng-route',
+    template: '<a *ngFor="let route of routes" routerLinkActive="router-link-active" [routerLink]="route.link">{{ route.text }}</a>'
+})
+export class RouteComponent {
+    protected routes: Array<RouteComponentData> = [];
+
+    setDynamicData(data: Array<RouteComponentData>): void {
+        data.forEach(item => {
+            this.routes.push(item);
+        });
+    }
+}
